Move settings functionality to dedicated package
